+++ app/src/main/java/hu/vmiklos/plees_tracker/AppDatabase.kt
@@ -13,9 +13,14 @@ import androidx.room.RoomDatabase
  * Contains the database holder and serves as the main access point for the
  * stored data.
  */
-@Database(entities = [Sleep::class], version = 4, exportSchema = false)
+@Database(
+    entities = [Sleep::class, HealthConnectDeletion::class],
+    version = 5,
+    exportSchema = false
+)
 abstract class AppDatabase : RoomDatabase() {
     abstract fun sleepDao(): SleepDao
+    abstract fun healthConnectDao(): HealthConnectDao
 }
 
 /* vim:set shiftwidth=4 softtabstop=4 expandtab: */